more gfa path tests
